Practice/API Scripts/game/websiteGame/requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
google-generativeai==0.8.3
google-cloud-secret-manager==2.22.0
gunicorn==23.0.0